requirements.txt
django
openai
faiss-cpu
pdfplumber
python-docx
django-cors-headers
drf-yasg
python-dotenv
